Thou shall not break the build (+ more warning fixes)

+++ Aquaria/AnimationEditor.h
@@ -94,7 +94,7 @@ public:
 
 	std::list<SkeletalSprite> undoHistory;
 
-	int undoEntry;
+	size_t undoEntry;
 
 	int currentKey;
